src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.6.15=h033912b_0
cycler=0.11.0=pypi_0
fonttools=4.37.1=pypi_0
kiwisolver=1.4.4=pypi_0
libffi=3.4.2=h0d85af4_5
libsqlite=3.39.2=h5a3d3bf_1
libzlib=1.2.12=hfe4f2af_2
matplotlib=3.5.3=pypi_0
ncurses=6.3=h96cf925_1
numpy=1.23.2=pypi_0
openssl=3.0.5=hb81d4ab_1
packaging=21.3=pypi_0
pillow=9.2.0=pypi_0
pip=22.2.2=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
python=3.9.13=hf8d34f4_0_cpython
python-dateutil=2.8.2=pypi_0
readline=8.1.2=h3899abd_0
setuptools=65.3.0=pyhd8ed1ab_1
six=1.16.0=pypi_0
sqlite=3.39.2=hd9f0692_1
tk=8.6.12=h5dbffcc_0
tzdata=2022c=h191b570_0
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.6=h775f41a_0
